requirements.txt
pandas
faker
psycopg2-binary
numpy
python-dotenv
python-dateutil
SQLAlchemy
python-decouple